Fixed segfault
Segfault was caused by not including -DGPU and -DCUDNN defines when
compiling the darknet_cpp files. The flags where also needed when
building an application on top of libdarknet_cpp.so.
Refactored detector module using pimpl idiom to avoid including darknet.h
in darknet.hpp. This removes the need for -DGPU and -DCUDNN for
compiling user applications on top of darknet_cpp.